# dist/api.yml
openapi: 3.0.3
info:
  version: '1.0.0'
  title: 'Web Development API'
  description: 'API for assignment 3.1'
...
paths:
  /Login:
    post:
      tags:
        - auth
      summary: Logs into an user given an ID and password.
      description: 'user logs in using username and password.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_login'
      responses:
        '200':
          description: 'successful log in.'
        '401':
          description: 'username or password incorrect.'
  /Logout:
    post:
      tags:
        - auth
      summary: User logs out from the system.
      responses:
        '200':
          description: 'successful log out.'
        '401':
          description: 'log out failed.'
  /Register:
    post:
      tags:
        - auth
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_register'
      responses:
        '201':
          description:  Added new user to the system.
        '400':
          description:  Wrong input.
...
  /matches:
    put:
      tags:
        - matches
      summary: Update match results.
      description: Get match ID and update values in match
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/update_match'
      responses:
        '201':
          description: 'Match updated.'
        '400':
          description: 'Wrong input.'
    post:
      tags:
        - matches
      summary: Create a match.
      description: Set a new match.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/new_match'
      responses:
        '201':
          description: 'created game'
        '404':
          description: 'match not found'       
  /matches/{match_id}:
    get:
      tags:
        - matches
      summary: Returns match details given it's ID.
      description: 'user provides match id and get all relevant info.'
      parameters:
        - name: match_id
          required: true
          in: path
          schema:
            type: integer
            example: 9
      responses:
        '200':
          description: 'successful fetched match info.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/game_details'
        '404':
          description: 'game data not found.'
    delete:
      tags:
        - matches
      summary: Delete a match.
      description: 'delete match from system.'
      parameters:
        - name: match_id
          required: true
          in: path
          schema:
            type: integer
      responses:
        '200':
          description: 'successful deleted match.'
        '404':
          description: 'match doesnt found.'      
  /matches/representative/get_all_matches:
    get:
      tags: 
        - matches
      summary: Get all existing matches.
      description: Get all existing matches.  
      responses:
        '200':
          description: 'successful retrieve'
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/get_all_matches'
        '500':
          description: 'something is wrong'
...
  /coaches/{coach_id}/preview:
    get:
      tags:
        - coaches
      summary: Returns coaches preview given it's ID.
      description: 'user provides coach id and get its preview info.'
      parameters:
        - name: coach_id
          required: true
          in: path
          schema:
            type: integer
            example: 459045
      responses:
        '200':
          description: 'successful fetched coach preview.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/coach_preview'
        '404':
          description: 'coach preview not found.'  
  /coaches/{coach_id}/full_data:
    get:
      tags:
        - coaches
      summary: Returns coaches full data given it's ID.
      description: 'user provides coach id and get its full info.'
      parameters:
        - name: coach_id
          required: true
          in: path
          schema:
            type: integer
            example: 459045
      responses:
        '200':
          description: 'successful fetched coach full data.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/coach_full_data'
        '404':
          description: 'player preview not found.'  
  /teams/{team_id}:
    get:
      tags:
      - teams
      summary: Return team data
      description: 'provide team id, get all players, history, etc.'
      parameters:
      - name: team_id
        required: true
        in: path
        schema:
          type: integer
          example: 85
      - name: include
        in: query
        schema:
          type: string
          example: matches
      responses:
        '200':
          description: 'successful fetched team.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/team_data'
        '404':
          description: 'team not found.'  
...
  /coaches/search:
    get:
      tags: 
        - coaches
      summary: Search for a coach
      description: 'find a coach given a string query.'
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
            example: Niels
      responses:
        '200':
          description: 'found matching results.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/coaches_ids'
        '404':
          description: 'no match found.'
  /teams/search:
    get:
      tags: 
        - teams
      summary: Search for a team
      description: 'find a team given a string query.'
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
            example: København
      responses:
        '200':
          description: 'found matching results.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/teams_ids'
        '404':
          description: 'no match found.'
components:
  schemas:
    get_all_matches:
      type: array
      items:
        type: integer
        example: 1
    user_login:
      type: object
      properties:
        username:
          type: string
          example: shiloari
        password:
          type: string
          example: g@1ThePro
    user_register:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 8
          description: 'using only letters'
          example: shiloari
          
        first_name:
          type: string
          example: Ariel
        
        last_name:
          type: string
          example: Shilo
          
        country:
          type: string
          example: Israel

        email:
          type: string
          example: qqq@qqq.com

        password:
          type: string
          minLength: 5
          maxLength: 10
          description: password must include at least 1 number and at least 1 special character.
          example: g@1ThePro
...
    game_details:
      type: object
      properties:
        Home_Team_ID:
          type: integer
          description: 'home team id'
          example: 85
        Away_Team_ID:
          type: integer
          description: 'away team id'
          example: 939
        Match_Date:
          type: string
          description: 'match date'
          example: '14/05/2021'
        Score:
          type: string
          description: 'match result (optional).'
          example: 1:2
        Stadium:
          type: string
          description: 'the stadium the match is played in.'
          example: Parken
        EventBook:
          description: 'event book: all events that happend in the game.'
          type: array
          items:
            type: string
            example: 04/06/20212 19:45 78, Goal - Stephan Maigaard Andersen
        Referee_ID:
          type: integer
          example: 1
...
    coach_preview:
      type: object
      properties:
        name:
          type: string
          description: 'coaches full name'
          example: Glen Riddersholm
        photo_path:
            type: integer
            description: 'photo path'
            example: 'https://cdn.sportmonks.com/images/soccer/players/5/459045.png'
        team_name:
            type: string
            description: 'team name'
            example: SønderjyskE
...
    coach_full_data:
        type: object
        properties:
          name:
            type: string
            description: 'coaches full name'
            example: Glen Riddersholm
          photo_path:
            type: string
            description: 'photo path'
            example: https://cdn.sportmonks.com/images/soccer/players/5/459045.png
          team_name:
            type: string
            description: 'team name'
            example: SønderjyskE
          common_name:
            type: string
            description: 'coach nickname'
            example: G. Riddersholm
          birthdate:
            type: string
            description: 'coaches birth date'
            example: 26/04/1972
          birth_country:
            type: string
            description: 'coaches origin country'
            example: Denmark
          nationality:
            type: string
            description: 'coaches nationality'
            example: Denmark
    update_match:
      type: object
      properties:
        match_id:
          type: integer
          description: "Match id to update."
          example: 9
        score:
          type: string
          description: match result (optional).
          example: '1:2'
        event_book:
          description: 'event book: all events that happend in the game.'
          type: array
          items:
            type: string
            example: 04/06/20212 19:45 78, Goal - Stephan Maigaard Andersen 
    new_match:
      type: object
      properties:
        home_team_id:
          type: integer
          description: home team id
          example: 85
        away_team_id:
          type: integer
          description: away team id
          example: 939
        date:
          type: string
          description: match date
          example: 14/05/2021
        hour:
          type: string
          description: match hour
          example: '20:00'
        referee_id:
          type: integer
          description: referee's id
          example: 1
    team_data:
      type: object
      properties:
        players_data:
          type: array
          items:
           type: object
           properties:
              name:
                type: string
                description: player name
                example: Stephan Maigaard Andersen
              image:
                type: string
                description: path to player's image
                example: https://cdn.sportmonks.com/images/soccer/players/2/26722.png
              position:
                type: integer
                description: player's position on field
                example: 1
              team_name:
                type: string
                description: team's name
                example: København      
        coach_data:
          description: coach's data.
          type: object
          properties:
            name:
              type: string
              example: Jess Thorup
            photo_path:
              type: string
              example: null
            team_name:
              type: string
              example: København           
        matches:
          type: array
          items:
            type: object
            properties:
              Match_ID:
                type: integer
                example: 2
              Home_Team_ID:
                type: integer
                example: 85
              Away_Team_ID:
                type: integer
                example: 939
              Match_Date:
                type: string
                example: 03/06/2021
              Stadium:
                type: string
                example: Parken
              Stage:
                type: integer
                example: Group Stage
              Score:
                type: string
                example: 3:1
              EventBook:
                type: array
                items:
                  type: string
                  example: 04/06/20212 19:45 78, Goal - Stephan Maigaard Andersen 
              Referee_ID:
                type: integer
                example: 2

      description: team's matches.
...
    coaches_ids:
      type: array
      items:
        properties:
          coach_id:
                type: integer
                description: coach id
                example: 457779
          coach_name:
                type: string
                description: Coach name
                example: Kent Nielsen           
    teams_ids:
      type: array
      items:
        properties:
          team_id:
                type: integer
                description: team id
                example: 3
          team_name:
                type: string
                description: team name
                example: 'Maccabi Haifa'